Create Cloud adoption Framework
Most of the DevSecOps principles and workflows follow a a sequence from Planning and Dev, all thru the SDLC process, to the post-prod/Operation stage. At the level of planning and development, the framework suggests 5 major areas or actions to look into. : Threat modelling, IDE security Plugins, pre-commit hooks, security coding standards and Code peer reviews. This file tends to start the conversation on that.